backend/requirements.txt
# FastAPI Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# OpenAI API Client (for LLM, TTS, STT)
openai==1.3.8

# Vector Database
chromadb==0.4.14

# Environment Variables Management
python-dotenv==1.0.0

# Image Processing (for screenshot compression)
pillow==10.1.0

# Screen Capture
pyautogui==0.9.53

# Async HTTP Client (optional, for better performance)
httpx==0.25.2
